First, ugly, stab at handling footnotes
Footnotes are implemented in the HTML as right-floated figures.
Take any of those and stuff them in LaTeX \footnote{} blocks. Also,
if the footnote starts with an asterisk, erase it from the footnote
and from the text.

The way this is done is horrible, risks erasing too much, doesn't
handle nested footnotes, and puts the footnote markers in the wrong
place - they are moved to the end of the paragraph. But doing it
properly requires going back and changing already transformed text.