Don't underline everything in a blockquote
Some authors user underlines for entire blockquotes, e.g. when including
letters read by the characters. This makes LaTeX mishandle line breaks
in the text.

Keep track of whether we're in a blockquote and ignore block underlines.
Underlined spans are still rendered - hopefully these will be smaller.

This seems to be enough to render correctly all (so far) 77 chapters of
https://www.fimfiction.net/story/432523/daily-equestria-life-with-monster-girl