Use fancy underlining and document that line breaks are manual
LaTeX really doesn't like underlined text, and none of the recommended
packages are complete solutions.

Default to a fancy style with the line above any descenders, broken
where they poke through.

Document that LaTeX will not automatically line break underlined text,
but print a list of the chapters where it occurs so the user can check
and fix it manually.