Let system() throw exceptions that kill the driver
When invoking xelatex, instead of having the system() call return nil
on errors and try to construct a new error message, let system() raise
an exception and let it out to the user. This is not any uglier, and
contains much more useful information.

E.g. not having "xelatex" in PATH goes from:
   Error running first render pass: pid 7536 exit 127 (RuntimeError)
to:
   No such file or directory - xelatex (Errno::ENOENT)