Add an option "-q" to transform ASCII quotes to Unicode
If all we have are ASCII (or HTML entity) quotation marks, LaTeX by
default renders them all as closing marks. This is not very pretty.
There is the "csquotes" package with the "\MakeOuterQuote{"}" command
which handles this situation automatically, but not if any quotation
marks are inside formatting, which they usually are in the source
documents.

Add an option "-q/--prettify-quotes" which blindly goes through the text
and changes every other ASCII quotation mark to alternating Unicode open
and close marks.