Clean up options processing
Make the options object an OpenStruct instead of a Hash, to reduce the
amount of brackets and colons in the code.

Collect default options in one location.

Keep the options object as a whole in the transformer instead of
splitting each key into its own local variable, since most of them are
only used once.