Treat the story ID as a string which is the working directory name
The story ID is always used as a string, so keep that as the data type
instead of converting to and from an integer.

Also remove the distinction between story IDs and working directories,
since it has never had a practical use, and the anthology transformer
requires that it does not exist.