Allow writing front matter to anthologies
This changes the way anthology authors and titles are handled, to
enable this information to be easily included in the front matter.
As a consequence, the configuration format is not compatible.

An anthology project created with an earlier version of FiMFic2PDF
will not work with this or later versions.