Print author and title in page headers for anthologies using novel
When producing an anthology using the "novel" class, use a page header
style that includes author and title on each page.

This unfortunately also places the page numbers in the header, but it's
what's available, and defining a totally custom header and footer style
turned out to be more trouble than it's worth.