Disable \nofiles command
In TeXLive 2024, the "novel" class runs the command \nofiles to disable
the generation of auxiliary output files. However, we need the .aux
file in order to read the document length from it.

Before setting the document class, override the definition of \nofiles
to be empty, so we will still get the .aux file.